fail gracefully if trying to delete nonexistent user
fixes https://sentry.io/organizations/coordinape/issues/2638695202/?project=5867811

+++ app/Http/Controllers/UserController.php
@@ -73,8 +73,10 @@ class UserController extends Controller
     }
 
     public function deleteUser(Request $request, $circle_id, $address) : JsonResponse  {
-
         $user = $request->user;
+        if(!$user)
+            return response()->json(['message' => 'User not found'], 422);
+
         $data = $this->repo->deleteUser($user);
         if(is_null($data)) {
             $error = ValidationException::withMessages([
